Will extract 'FeedStoreSpy' into a shared scope to remove duplication
