+++ ModularFeedTDDCaseStudy/ModularFeedTDDCaseStudyTests/FeedCache/Helpers/FeedStoreSpy.swift
@@ -1,0 +1,48 @@
+//
+//  FeedStoreSpy.swift
+//  ModularFeedTDDCaseStudyTests
+//
+//  Created by Clint M on 3/23/19.
+//  Copyright © 2019 Clint M. All rights reserved.
+//
+
+import Foundation
+import ModularFeedTDDCaseStudy
+
+class FeedStoreSpy: FeedStore {
+    enum ReceivedMessage: Equatable {
+        case deleteCacheFeed
+        case insert([LocalFeedImage], Date)
+    }
+    
+    private(set) var receivedMessages = [ReceivedMessage]()
+    
+    private var deletionCompletions = [DeletionCompletion]()
+    private var insertionCompletions = [InsertionCompletion]()
+    
+    func deleteCachedFeed(completion: @escaping DeletionCompletion) {
+        deletionCompletions.append(completion)
+        receivedMessages.append(.deleteCacheFeed)
+    }
+    
+    func completeDeletion(with error: Error, at index: Int = 0) {
+        deletionCompletions[index](error)
+    }
+    
+    func completeDeletionSuccessfully(at index: Int = 0) {
+        deletionCompletions[index](nil)
+    }
+    
+    func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
+        insertionCompletions.append(completion)
+        receivedMessages.append(.insert(feed, timestamp))
+    }
+    
+    func completeInsertion(with error: Error, at index: Int = 0) {
+        insertionCompletions[index](error)
+    }
+    
+    func completeInsertionSuccessfully(at index: Int = 0) {
+        insertionCompletions[index](nil)
+    }
+}
